expose sample generation logic
