+++ .gitignore
@@ -1,5 +1,6 @@
 # Folders
 .idea
+.vscode
 
 # Build artifacts
 /build
