+++ hack/generate/samples/ansible/constants.go
