+++ hack/generate/samples/ansible/testdata/ansible.cfg
