+++ hack/generate/samples/ansible/testdata/fixture_collection/galaxy.yml
