+++ hack/generate/samples/ansible/testdata/fixture_collection/roles/dummy/tasks/main.yml
