+++ hack/generate/samples/ansible/testdata/inventory/group_vars/test.yml
