+++ hack/generate/samples/ansible/testdata/inventory/hosts
