+++ hack/generate/samples/ansible/testdata/plugins/filter/test.py
