+++ hack/generate/samples/ansible/testdata/secret.yml
