+++ hack/generate/samples/ansible/testdata/tasks/argstest_test.yml
