+++ hack/generate/samples/ansible/testdata/tasks/casetest_test.yml
