+++ hack/generate/samples/ansible/testdata/tasks/clusterannotationtest_test.yml
