+++ hack/generate/samples/ansible/testdata/tasks/collectiontest_test.yml
