+++ hack/generate/samples/ansible/testdata/tasks/inventorytest_test.yml
