+++ hack/generate/samples/ansible/testdata/tasks/secretstest_test.yml
