+++ hack/generate/samples/ansible/testdata/tasks/selectortest_test.yml
