+++ hack/generate/samples/ansible/testdata/tasks/subresourcestest_test.yml
