+++ hack/generate/samples/ansible/testdata/watches.yaml
