+++ hack/generate/samples/helm/testdata/memcached-0.0.1.tgz
