+++ hack/generate/samples/pkg/context.go
