+++ hack/generate/samples/pkg/utils.go
