+++ test/e2e/go/local_test.go
@@ -18,31 +18,32 @@ import (
 	"os/exec"
 
 	. "github.com/onsi/ginkgo" //nolint:golint
-	. "github.com/onsi/gomega" //nolint:golint
+	. "github.com/onsi/gomega"
+	"github.com/operator-framework/operator-sdk/testutils/e2e/operator"
 )
 
 var _ = Describe("Running Go projects", func() {
 	Context("built with operator-sdk", func() {
-
 		BeforeEach(func() {
-			By("installing CRD's")
-			err := tc.Make("install")
+			By("Installing CRD's")
+			err := operator.InstallCRDs(goSample)
 			Expect(err).NotTo(HaveOccurred())
 		})
 
 		AfterEach(func() {
-			By("uninstalling CRD's")
-			err := tc.Make("uninstall")
+			By("Uninstalling CRD's")
+			err := operator.UninstallCRDs(goSample)
 			Expect(err).NotTo(HaveOccurred())
 		})
 
-		It("should run correctly locally", func() {
-			By("running the project")
+		It("Should run correctly when run locally", func() {
+			By("Running the project")
 			cmd := exec.Command("make", "run")
+			cmd.Dir = goSample.Dir()
 			err := cmd.Start()
 			Expect(err).NotTo(HaveOccurred())
 
-			By("killing the project")
+			By("Killing the project")
 			err = cmd.Process.Kill()
 			Expect(err).NotTo(HaveOccurred())
 		})
